Renamed modeling.py to features.py
